refactor: move statics out of Wall
