+++ src/main/java/org/runejs/client/Class13.java
@@ -10,7 +10,6 @@ import org.runejs.client.media.renderable.actor.Actor;
 import org.runejs.client.media.renderable.actor.Player;
 import org.runejs.client.media.renderable.actor.PlayerAppearance;
 import org.runejs.client.scene.GroundItemTile;
-import org.runejs.client.scene.tile.Wall;
 
 public class Class13 {
     public static boolean[] playerArray = new boolean[5];
@@ -21,6 +20,7 @@ public class Class13 {
      */
     public static IndexedImage tabHighlightImageBottomRight;
     public static int[] objectDataIds;
+    private static int anInt356 = 0;
 
     public static void handleActorAnimation(Actor actor) {
         if(actor.worldX < 128 || actor.worldY < 128 || actor.worldX >= 13184 || actor.worldY >= 13184) {
@@ -82,9 +82,9 @@ public class Class13 {
     }
 
     public static synchronized byte[] method246(int arg1) {
-        if(arg1 == 100 && Wall.anInt356 > 0) {
-            byte[] is = GroundItemTile.aByteArrayArray1377[--Wall.anInt356];
-            GroundItemTile.aByteArrayArray1377[Wall.anInt356] = null;
+        if(arg1 == 100 && anInt356 > 0) {
+            byte[] is = GroundItemTile.aByteArrayArray1377[--anInt356];
+            GroundItemTile.aByteArrayArray1377[anInt356] = null;
             return is;
         }
         if(arg1 == 5000 && Main.anInt1764 > 0) {
