+++ src/main/java/org/runejs/client/cache/media/gameInterface/GameInterface.java
@@ -33,7 +33,6 @@ import org.runejs.client.node.CachedNode;
 import org.runejs.client.scene.InteractiveObject;
 import org.runejs.client.scene.SceneCluster;
 import org.runejs.client.scene.tile.GenericTile;
-import org.runejs.client.scene.tile.Wall;
 import org.runejs.client.scene.tile.WallDecoration;
 import org.runejs.client.scene.util.CollisionMap;
 import org.runejs.client.util.Signlink;
@@ -61,6 +60,7 @@ public class GameInterface extends CachedNode {
      */
     public static IndexedImage tabHighlightImageTopLeft;
     public static int anInt1233 = 0;
+    public static GameInterface aGameInterface_353 = null;
 
     public boolean isNewInterfaceFormat = false;
     public int contentType;
@@ -1560,8 +1560,8 @@ public class GameInterface extends CachedNode {
 
         runClientScriptsForInterface(arg5, arg3, 0, arg2, arg6, -1, cachedInterfaces[widgetId], arg1, 0);
 
-        if(Wall.aGameInterface_353 != null) {
-            GameInterface gameInterface = Wall.aGameInterface_353;
+        if(aGameInterface_353 != null) {
+            GameInterface gameInterface = aGameInterface_353;
             GameInterface gameInterface_24_ = FramemapDefinition.method878(gameInterface);
             if(gameInterface_24_ != null) {
                 int[] is = Class13.method247(gameInterface_24_);
@@ -1580,14 +1580,14 @@ public class GameInterface extends CachedNode {
                 if(gameInterface_24_.originalWidth < gameInterface.originalWidth + i_26_) {
                     i_26_ = gameInterface_24_.originalWidth - gameInterface.originalWidth;
                 }
-                if(Wall.aGameInterface_353.anObjectArray2669 != null && (arg2 & 0x200) != 0) {
+                if(aGameInterface_353.anObjectArray2669 != null && (arg2 & 0x200) != 0) {
                     ClientScriptRunner.runClientScripts(gameInterface.anObjectArray2669, 0, i, gameInterface, i_26_);
                 }
                 if(MouseHandler.currentMouseButtonPressed == 0 && (arg2 & 0x400) != 0) {
-                    if(Wall.aGameInterface_353.anObjectArray2695 != null) {
+                    if(aGameInterface_353.anObjectArray2695 != null) {
                         ClientScriptRunner.runClientScripts(gameInterface.anObjectArray2695, 0, i, gameInterface, i_26_);
                     }
-                    Wall.aGameInterface_353 = null;
+                    aGameInterface_353 = null;
                 }
             }
         }
@@ -1767,14 +1767,14 @@ public class GameInterface extends CachedNode {
                     if (MouseHandler.currentMouseButtonPressed == 1 && bool)
                         bool_9_ = true;
                     boolean bool_10_ = false;
-                    if (gameInterface.anInt2738 != -1 && bool_9_ && Wall.aGameInterface_353 == null) {
+                    if (gameInterface.anInt2738 != -1 && bool_9_ && aGameInterface_353 == null) {
                         MovedStatics.anInt1996 = MouseHandler.mouseX;
-                        Wall.aGameInterface_353 = gameInterface;
+                        aGameInterface_353 = gameInterface;
                         MovedStatics.anInt2621 = MouseHandler.mouseY;
                     }
                     if (MouseHandler.clickType == 1 && i_6_ <= MouseHandler.clickX && MouseHandler.clickY >= i_4_ && MouseHandler.clickX < i_8_ && MouseHandler.clickY < i_7_)
                         bool_10_ = true;
-                    if (Wall.aGameInterface_353 != null) {
+                    if (aGameInterface_353 != null) {
                         bool_9_ = false;
                         bool = false;
                         bool_10_ = false;
